Finished Problem 1.0
The problem is finished and I have rewrote the code to be used for general cases! This can be set to public now as well since personal information has been removed. I kept both versions, one that was turned into coursera and was tailored for linux and the other that is tailored general case for any os.