Turn in Commit
The project has been turned in and the summary pdf created has been included.

I also added the two modules the lab wanted me to use instead of crafting the functions myself, these are emails.py and reports.py.

I also added the complete code that was modified for the qwiklabs server in workingground and of course in final.